add DietPlan

+++ Westeros.Diet.Data/Model/DietPlan.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public int Calories { get; set; }
-        public decimal Proteins { get; set; }
-        public decimal Carbohydrates { get; set; }
-        public decimal Fats { get; set; }
+        public double Proteins { get; set; }
+        public double Carbohydrates { get; set; }
+        public double Fats { get; set; }
         public UserProfile UserProfile { get; set; }
     }
 }
